refactor all packages into internal/ or pkg/
