+++ cmd/waddles/main.go
@@ -1,7 +1,9 @@
 package main
 
-import "github.com/the-sanctuary/waddles"
+import "github.com/the-sanctuary/waddles/internal/waddles"
 
 func main() {
-	waddles.Run()
+	wadl := new(waddles.Waddles)
+
+	wadl.Run()
 }
