+++ internal/model/nitro_user_channel.go
