+++ pkg/command/command_test.go
