+++ pkg/command/router_test.go
